refactor: `FieldTypes` and `QuestionTypes` to `enums.py`
